backend/requirements.txt
django>=2.1.2
psycopg2==2.7.5
djangorestframework==3.9.0
django-cors-headers==2.4.0
django-extensions==2.1.4
coreapi==2.3.3
pyyaml=3.13

